Removed "Update Command" as it is not needed also changed some return types of the CommandController

Took 2 minutes
